fix the issues from the review
